Sentiment Analysis
sentiment_analysis_functions.py ==> Functions to perform antonym & negation
Unit_tests ++> Unit test cases for antonym & negation functions
requirements.txt ==> Dependencies for venv
IMDB_Data ==> Source dataset for training
